Gyro measurement for the Raspberry Pi with the MPU 6050
The python files feature a (currently) not object oriented way of
automatically starting a gyroscope measurement by pressing a button
(please see the schematics in the pdf). The measurement is handeled with
the gyro_normal.py or gyro_filter.py files (the one just takes the
calibrated raw data, the other one filters the raw data).